restructure every page to import from client components
